+++ apps/web/src/components/elementId.ts
@@ -1,0 +1,4 @@
+export const elementIds = {
+  bodyTitle: "body-title",
+  scrollableContainer: "scrollable-container",
+}