+++ apps/web/src/components/errorBoundary.tsx
@@ -22,7 +22,7 @@ export class ErrorBoundary extends React.Component<Props, State> {
   render() {
     if (this.state.hasError) {
       return (
-        <div className="flex h-screen w-screen flex-col items-center justify-center gap-4">
+        <div className="flex h-dvh w-screen flex-col items-center justify-center gap-4">
           <div className="text-4xl font-bold">{L.errorPage.serverError}</div>
           <a href="/" className="text-link hover:text-link-hover text-lg underline transition-colors">
             {L.common.goBackToHome}