+++ apps/web/src/components/layout/body.tsx
@@ -2,6 +2,7 @@ import clsx from "clsx"
 import { ReactElement, forwardRef, useEffect, useRef, useState } from "react"
 
 import { ContainerSizeProvider } from "../../contexts/containerSize"
+import { elementIds } from "../elementId"
 
 interface BodyProps extends React.HTMLAttributes<HTMLDivElement> {
   title?: string
@@ -13,7 +14,7 @@ interface BodyProps extends React.HTMLAttributes<HTMLDivElement> {
 const Title = forwardRef<HTMLDivElement, { title: string; description?: string | ReactElement; tail?: ReactElement }>(
   ({ title, description, tail }: { title: string; description?: string | ReactElement; tail?: ReactElement }, ref) => {
     return (
-      <div className="mt-8" ref={ref}>
+      <div className="mt-8" ref={ref} id={elementIds.bodyTitle}>
         <div className="flex h-10 items-center justify-between text-2xl font-bold">
           <h1 className="flex-1">{title}</h1>
           <div className="flex-0">{tail}</div>