+++ apps/web/src/components/layout/page.tsx
@@ -1,9 +1,8 @@
 import clsx, { ClassValue } from "clsx"
 import { ComponentPropsWithoutRef } from "react"
 
+import { elementIds } from "../elementId"
 import { SideMenu } from "../sidemenu"
-
-export const ScrollableContainerId = "scrollable-container"
 
 type PageLayoutProps = ComponentPropsWithoutRef<"div"> & {
   containerClassName?: ClassValue
@@ -11,11 +10,11 @@ type PageLayoutProps = ComponentPropsWithoutRef<"div"> & {
 
 export const PageLayout = ({ children, containerClassName, ...props }: PageLayoutProps) => {
   return (
-    <div className="flex h-screen w-screen" {...props}>
+    <div className="flex h-dvh w-screen" {...props}>
       <SideMenu />
       <div className="w-full">
         {/* <div className="bg-card flex h-16 w-full justify-end border-b"></div> */}
-        <div id={ScrollableContainerId} className={clsx("h-[calc(100vh)] w-full overflow-y-auto", "pb-8", containerClassName)}>
+        <div id={elementIds.scrollableContainer} className={clsx("h-dvh w-full overflow-y-auto", "pb-8", containerClassName)}>
           {children}
         </div>
       </div>